prim should live with not great printing of class
